[#1] You no longer double back in snake game.
This was implemented with two variables, a current and next direction.
Re-using the current direction meant that the opposite direction kept being evaluated on the current direction before movement had taken place. This lead to an effect where the update loop did not double back but because the movement loop is slower, it actually did.